refactor(): * extract lieFlat & its sub functions * added a lot of docstrings * various cleanups all over
