+++ src/utils/csgProjections.js
@@ -7,15 +7,15 @@ const {EPS} = require('../constants')
 const projectToOrthoNormalBasis = function (csg, orthobasis) {
   let cags = []
   csg.polygons.filter(function (p) {
-                // only return polys in plane, others may disturb result
+    // only return polys in plane, others may disturb result
     return p.plane.normal.minus(orthobasis.plane.normal).lengthSquared() < (EPS * EPS)
   })
-            .map(function (polygon) {
-              let cag = polygon.projectToOrthoNormalBasis(orthobasis)
-              if (cag.sides.length > 0) {
-                cags.push(cag)
-              }
-            })
+  .map(function (polygon) {
+    let cag = polygon.projectToOrthoNormalBasis(orthobasis)
+    if (cag.sides.length > 0) {
+      cags.push(cag)
+    }
+  })
   let result = new CAG().union(cags)
   return result
 }
